fix: 검색 페이지 옵션 팝업 에러 해결
