+++ src/component/Layout/Layout.tsx
@@ -1,24 +1,11 @@
-import { Outlet, useLocation, useNavigate } from 'react-router-dom'
+import { Outlet, useLocation } from 'react-router-dom'
 import styled from 'styled-components'
-import { useQuery } from '@tanstack/react-query'
 import Header from '../header/header-main/Header'
 import { ROUTER_PATH, ROUTER_PATH_REGEX } from '@/constants/constant'
 import HeaderSub from '../header/header-sub/HeaderSub'
 
-import { useSearchModalFullStore } from '@/store/useSearchModalFullStore'
-
-import { CiTimer } from 'react-icons/ci'
-import * as S from './Layout.module'
-import { GoArrowUpLeft } from 'react-icons/go'
-import { getSearch } from '@/apis/search'
-
-import {
-  useSearchTermListStore,
-  useSearchTermStore
-} from '@/store/useSearchTermStore'
-import { IPlayList } from '@/types/playList'
-import { useDebounce } from '@/hooks/useDebounce'
-import { Navbar, ModalFull } from '@/component'
+import { Navbar } from '@/component'
+import SearchList from '../Search/SearchList'
 
 const Container = styled.div`
   max-width: var(--max-width);
@@ -57,13 +44,6 @@ const Layout = () => {
 
   const isSubHeaderPaths = ''
 
-  const modalFullOpen = useSearchModalFullStore(state => state.state)
-  const setModalFull = useSearchModalFullStore(state => state.setModalState)
-  const searchTerm = useSearchTermStore(state => state.searchTerm)
-  const setSearchTerm = useSearchTermStore(state => state.setSearchTerm)
-  const addSearchTerm = useSearchTermListStore(state => state.addSearchTerm)
-  const searchTermList = useSearchTermListStore(state => state.searchTermList)
-  const navigate = useNavigate()
   const renderHeader = () => {
     if (isSubHeaderPaths) {
       return <HeaderSub />
@@ -76,30 +56,6 @@ const Layout = () => {
     }
   }
 
-  const debouncedTitle = useDebounce(searchTerm, 300)
-
-  const { data: searchs, isError } = useQuery<IPlayList[]>({
-    queryKey: ['searchPlayList', debouncedTitle],
-    queryFn: () => getSearch(debouncedTitle),
-    enabled: !!debouncedTitle,
-    staleTime: 1000 * 60
-  })
-
-  const handleMoalFullClose = () => {
-    setModalFull(false)
-    setSearchTerm('')
-  }
-
-  const handleSearchClick = (term?: string) => {
-    setModalFull(false) // 모달 닫기
-    const searchValue = term || searchTerm // term이 있을 경우 사용, 아니면 searchTerm 사용
-    setSearchTerm(searchValue) // 선택된 검색어 설정
-    addSearchTerm(searchValue) // 최근 검색어 목록에 추가
-    navigate(`/search?search=${searchValue}`, { state: searchValue }) // navigate로 검색어와 함께 이동
-  }
-
-  if (isError) throw new Error('에러가 발생했습니다.')
-
   return (
     <>
       <Container className={!isNoNavbarPaths ? 'is-navbar' : ''}>
@@ -107,43 +63,7 @@ const Layout = () => {
         <Outlet />
         {!isNoNavbarPaths && <Navbar />}
       </Container>
-      <ModalFull
-        id={'2'}
-        closeModal={handleMoalFullClose}
-        isOpen={modalFullOpen}>
-        {searchTerm.length > 0
-          ? searchs?.map((search: IPlayList) => (
-              <S.ModalfullContent
-                key={search.playlist_id}
-                onClick={() => handleSearchClick()}>
-                <S.ModalFullTop>
-                  <CiTimer onClick={() => handleSearchClick()} />
-                </S.ModalFullTop>
-                <S.ModalFullMid>
-                  <S.ModalSpan>{search.title}</S.ModalSpan>
-                </S.ModalFullMid>
-                <S.ModalFullBottom>
-                  <GoArrowUpLeft />
-                </S.ModalFullBottom>
-              </S.ModalfullContent>
-            ))
-          : searchTermList &&
-            searchTermList?.map(search => (
-              <S.ModalfullContent
-                key={search.id}
-                onClick={() => handleSearchClick(search.term)}>
-                <S.ModalFullTop>
-                  <CiTimer />
-                </S.ModalFullTop>
-                <S.ModalFullMid>
-                  <S.ModalSpan>{search.term}</S.ModalSpan>
-                </S.ModalFullMid>
-                <S.ModalFullBottom>
-                  <GoArrowUpLeft />
-                </S.ModalFullBottom>
-              </S.ModalfullContent>
-            ))}
-      </ModalFull>
+      <SearchList />
     </>
   )
 }
