+++ src/pages/search/index.tsx
@@ -1,6 +1,6 @@
 import { getSearchPlayLists } from '@/apis/search/playList/index'
 import { useInfiniteQuery } from '@tanstack/react-query'
-import { useRef } from 'react'
+import { useEffect, useRef, useState } from 'react'
 import * as S from '@/component/FeedList/FeedList.style'
 import { FeedList, PlayListSkeleton } from '@/component'
 import { IPlayListType } from '@/types/playList'
@@ -10,7 +10,9 @@ import { useLocation } from 'react-router-dom'
 export const SearchPage = () => {
   const location = useLocation()
 
-  const searchTerm = location.state
+  const searchUrl = location.state
+
+  const [searchTerm, setSearchTerm] = useState<string>(searchUrl)
 
   const observerElem = useRef<HTMLDivElement | null>(null)
   const {
@@ -45,6 +47,12 @@ export const SearchPage = () => {
     isFetchingNextPage,
     fetchNextPage
   })
+
+  useEffect(() => {
+    if (location.state && location.state.searchTerm) {
+      setSearchTerm(location.state.searchTerm)
+    }
+  }, [location.state])
 
   if (isError)
     return (
